Add files via upload
This is the YOLOv7 in works from site: https://github.com/Chris-hughes10/Yolov7-training/blob/main/blog%20post/Yolo7%20blog%20post.ipynb. 
I am currently having troubles with albumentation and getting the axes to be the same.  You will find the errors in the file.  Finally, this allows you download a similar dataset of knees and use as an example.  The dataset is small.  I have not broken functions up as of now, so it is one large file with many functions.